requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
sentence-transformers
pandas
fpdf
